requirements.txt
static-ffmpeg 
pydub 
nemo_toolkit['asr']

typer[all]
fastapi
uvicorn[standard]
python-multipart
